requirements.txt
# ==================== Trading AI Bot - Backend Requirements ====================
# تم إنشاء هذا الملف: 2025-11-20
# ⚠️ تحذير: لا تحذف هذا الملف - إنه ضروري لتشغيل النظام
# 
# لتثبيت جميع المكتبات:
#   pip install -r requirements.txt
#
# لتثبيت مكتبة واحدة:
#   pip install package_name==version
#
# لتحديث جميع المكتبات:
#   pip install --upgrade -r requirements.txt

# ==================== FastAPI & Web Framework ====================
fastapi==0.104.1
uvicorn[standard]==0.24.0
flask==3.0.0
flask-cors==4.0.0
python-multipart==0.0.6

# ==================== Database ====================
# sqlite3  # Built-in with Python (no need to install)
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.10

# ==================== Authentication & Security ====================
pyjwt==2.10.1
bcrypt==4.1.1
cryptography==41.0.7
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.3.0
dnspython==2.8.0

# ==================== API & HTTP ====================
requests==2.31.0
httpx>=0.24.0
aiohttp==3.9.1

# ==================== Binance Trading ====================
python-binance==1.0.17
ccxt==4.0.96

# ==================== Data Processing & Analysis ====================
pandas>=2.1.3
numpy>=1.26.2
scipy==1.11.4
scikit-learn==1.3.2

# ==================== Machine Learning ====================
xgboost==2.0.3
joblib==1.3.2

# ==================== Rate Limiting & Performance ====================
slowapi==0.1.9
Flask-Limiter==3.5.0
redis==5.0.1
cachetools==5.3.2

# ==================== Logging & Monitoring ====================
python-json-logger==2.0.7
sentry-sdk==1.38.0
prometheus-client==0.19.0
psutil==5.9.6

# ==================== Task Scheduling ====================
schedule==1.2.0
APScheduler==3.10.4
celery==5.3.4

# ==================== Email & Notifications ====================
python-telegram-bot>=20.3
firebase-admin==6.2.0

# ==================== Testing ====================
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# ==================== Development Tools ====================
black==23.12.0
flake8==6.1.0
pylint==3.0.3
isort==5.13.2
mypy==1.7.1

# ==================== Utilities ====================
python-dateutil==2.8.2
pytz==2023.3
click==8.1.7
tqdm==4.66.1
colorama==0.4.6
tabulate==0.9.0

# ==================== Optional: Advanced Features ====================
# Uncomment if needed:
# tensorflow==2.15.0  # For ML models
# torch==2.1.1  # For deep learning
# plotly==5.18.0  # For advanced charts
# dash==2.14.1  # For dashboards
